Move operand resolution into `table_slice.hpp`
