+++ libvast/src/table_slice.cpp
@@ -718,4 +718,109 @@ table_slice resolve_enumerations(table_slice slice) {
   };
 }
 
+auto resolve_meta_extractor(const table_slice& slice, const meta_extractor& ex)
+  -> view<data> {
+  if (slice.encoding() == table_slice_encoding::none)
+    return {};
+  switch (ex.kind) {
+    case meta_extractor::type: {
+      return slice.schema().name();
+    }
+    case meta_extractor::import_time: {
+      const auto import_time = slice.import_time();
+      if (import_time == time{}) {
+        // The table slice API returns the epoch timestamp if no import
+        // time was set, so we convert that to null.
+        return {};
+      }
+      return import_time;
+    }
+  }
+  die("unhandled meta extractor kind");
+}
+
+auto resolve_operand(const table_slice& slice, const operand& op)
+  -> std::pair<type, std::shared_ptr<arrow::Array>> {
+  if (slice.encoding() == table_slice_encoding::none)
+    return {};
+  const auto batch = to_record_batch(slice);
+  const auto& layout = caf::get<record_type>(slice.schema());
+  auto inferred_type = type{};
+  auto array = std::shared_ptr<arrow::Array>{};
+  // Helper function that binds a fixed value.
+  auto bind_value = [&](const data& value) {
+    inferred_type = type::infer(value);
+    if (not inferred_type) {
+      inferred_type = type{string_type{}};
+      // VAST has no N/A type equivalent for Arrow, so we just use a string type
+      // here.
+      auto builder
+        = string_type::make_arrow_builder(arrow::default_memory_pool());
+      const auto append_result = builder->AppendNulls(batch->num_rows());
+      VAST_ASSERT(append_result.ok(), append_result.ToString().c_str());
+      array = builder->Finish().ValueOrDie();
+      return;
+    }
+    auto g = [&]<concrete_type Type>(const Type& inferred_type) {
+      auto builder
+        = inferred_type.make_arrow_builder(arrow::default_memory_pool());
+      for (int i = 0; i < batch->num_rows(); ++i) {
+        const auto append_result
+          = append_builder(inferred_type, *builder,
+                           make_view(caf::get<type_to_data_t<Type>>(value)));
+        VAST_ASSERT(append_result.ok(), append_result.ToString().c_str());
+      }
+      array = builder->Finish().ValueOrDie();
+    };
+    caf::visit(g, inferred_type);
+  };
+  // Helper function that binds an existing array.
+  auto bind_array = [&](const offset& index) {
+    inferred_type = layout.field(index).type;
+    array = static_cast<arrow::FieldPath>(index).Get(*batch).ValueOrDie();
+  };
+  auto f = detail::overload{
+    [&](const data& value) {
+      bind_value(value);
+    },
+    [&](const field_extractor& ex) {
+      for (const auto& index :
+           layout.resolve_key_suffix(ex.field, slice.schema().name())) {
+        bind_array(index);
+        return;
+      }
+      bind_value({});
+    },
+    [&](const type_extractor& ex) {
+      for (const auto& [field, index] : layout.leaves()) {
+        bool match = field.type == ex.type;
+        if (not match) {
+          for (auto name : field.type.names()) {
+            if (name == ex.type.name()) {
+              match = true;
+              break;
+            }
+          }
+        }
+        if (match) {
+          bind_array(index);
+          return;
+        }
+      }
+      bind_value({});
+    },
+    [&](const meta_extractor& ex) {
+      bind_value(materialize(resolve_meta_extractor(slice, ex)));
+    },
+    [&](const data_extractor& ex) {
+      bind_array(layout.resolve_flat_index(ex.column));
+    },
+  };
+  caf::visit(f, op);
+  return {
+    std::move(inferred_type),
+    std::move(array),
+  };
+}
+
 } // namespace vast
